fix: env in payload
send the environment to css app
