+++ docker/kc-cron-job/.dockerignore
@@ -2,3 +2,4 @@ node_modules
 npm-debug.log
 __tests__
 jest.config.js
+.env
